Separate RentaiService from VRUI
