+++ VRUI.java
@@ -6,205 +6,25 @@ import java.util.Scanner;
 public class VRUI {
 	private static Scanner scanner = new Scanner(System.in) ;
 
-	private List<Customer> customers = new ArrayList<Customer>() ;
-
-	private List<Video> videos = new ArrayList<Video>() ;
+	private RentalService rentalService = new RentalService();
 
 	public static void main(String[] args) {
-		VRUI ui = new VRUI() ;
+		VRUI ui = new VRUI();
 
 		boolean quit = false ;
 		while ( ! quit ) {
-			int command = ui.showCommand() ;
-			switch ( command ) {
-				case 0: quit = true ; break ;
-				case 1: ui.listCustomers() ; break ;
-				case 2: ui.listVideos() ; break ;
-				case 3: ui.register("customer") ; break ;
-				case 4: ui.register("video") ; break ;
-				case 5: ui.rentVideo() ; break ;
-				case 6: ui.returnVideo() ; break ;
-				case 7: ui.getCustomerReport() ; break;
-				case 8: ui.clearRentals() ; break ;
-				case -1: ui.init() ; break ;
-				default: break ;
+			ui.showCommand();
+			int command = ui.inputCommand();
+			quit = ui.isQuit(command);
+			if(!quit) {
+				ui.processCommand(command);
 			}
 		}
 		System.out.println("Bye");
 	}
 
-	public void clearRentals() {
-		System.out.println("Enter customer name: ") ;
-		String customerName = scanner.next() ;
 
-		Customer foundCustomer = null ;
-		for ( Customer customer: customers ) {
-			if ( customer.getName().equals(customerName)) {
-				foundCustomer = customer ;
-				break ;
-			}
-		}
-
-		if ( foundCustomer == null ) {
-			System.out.println("No customer found") ;
-		} else {
-			System.out.println("Name: " + foundCustomer.getName() +
-					"\tRentals: " + foundCustomer.getRentals().size()) ;
-			for ( Rental rental: foundCustomer.getRentals() ) {
-				System.out.print("\tTitle: " + rental.getVideo().getTitle() + " ") ;
-				System.out.print("\tPrice Code: " + rental.getVideo().getPriceCode()) ;
-			}
-
-			List<Rental> rentals = new ArrayList<Rental>() ;
-			foundCustomer.setRentals(rentals);
-		}
-	}
-
-	public void returnVideo() {
-		System.out.println("Enter customer name: ") ;
-		String customerName = scanner.next() ;
-
-		Customer foundCustomer = null ;
-		for ( Customer customer: customers ) {
-			if ( customer.getName().equals(customerName)) {
-				foundCustomer = customer ;
-				break ;
-			}
-		}
-		if ( foundCustomer == null ) return ;
-
-		System.out.println("Enter video title to return: ") ;
-		String videoTitle = scanner.next() ;
-
-		List<Rental> customerRentals = foundCustomer.getRentals() ;
-		for ( Rental rental: customerRentals ) {
-			if ( rental.getVideo().getTitle().equals(videoTitle) && rental.getVideo().isRented() ) {
-				rental.returnVideo();
-				rental.getVideo().setRented(false);
-				break ;
-			}
-		}
-	}
-
-	private void init() {
-		Customer james = new Customer("James") ;
-		Customer brown = new Customer("Brown") ;
-		customers.add(james) ;
-		customers.add(brown) ;
-
-		Video v1 = new Video("v1", Video.CD, Video.REGULAR, new Date()) ;
-		Video v2 = new Video("v2", Video.DVD, Video.NEW_RELEASE, new Date()) ;
-		videos.add(v1) ;
-		videos.add(v2) ;
-
-		Rental r1 = new Rental(v1) ;
-		Rental r2 = new Rental(v2) ;
-
-		james.addRental(r1) ;
-		james.addRental(r2) ;
-	}
-
-	public void listVideos() {
-		System.out.println("List of videos");
-
-		for ( Video video: videos ) {
-			System.out.println("Price code: " + video.getPriceCode() +"\tTitle: " + video.getTitle()) ;
-		}
-		System.out.println("End of list");
-	}
-
-	public void listCustomers() {
-		System.out.println("List of customers");
-		for ( Customer customer: customers ) {
-			System.out.println("Name: " + customer.getName() +
-					"\tRentals: " + customer.getRentals().size()) ;
-			for ( Rental rental: customer.getRentals() ) {
-				System.out.print("\tTitle: " + rental.getVideo().getTitle() + " ") ;
-				System.out.print("\tPrice Code: " + rental.getVideo().getPriceCode()) ;
-			}
-		}
-		System.out.println("End of list");
-	}
-
-	public void getCustomerReport() {
-		System.out.println("Enter customer name: ") ;
-		String customerName = scanner.next() ;
-
-		Customer foundCustomer = null ;
-		for ( Customer customer: customers ) {
-			if ( customer.getName().equals(customerName)) {
-				foundCustomer = customer ;
-				break ;
-			}
-		}
-
-		if ( foundCustomer == null ) {
-			System.out.println("No customer found") ;
-		} else {
-			String result = foundCustomer.getReport() ;
-			System.out.println(result);
-		}
-	}
-
-	public void rentVideo() {
-		System.out.println("Enter customer name: ") ;
-		String customerName = scanner.next() ;
-
-		Customer foundCustomer = null ;
-		for ( Customer customer: customers ) {
-			if ( customer.getName().equals(customerName)) {
-				foundCustomer = customer ;
-				break ;
-			}
-		}
-
-		if ( foundCustomer == null ) return ;
-
-		System.out.println("Enter video title to rent: ") ;
-		String videoTitle = scanner.next() ;
-
-		Video foundVideo = null ;
-		for ( Video video: videos ) {
-			if ( video.getTitle().equals(videoTitle) && video.isRented() == false ) {
-				foundVideo = video ;
-				break ;
-			}
-		}
-
-		if ( foundVideo == null ) return ;
-
-		Rental rental = new Rental(foundVideo) ;
-		foundVideo.setRented(true);
-
-		List<Rental> customerRentals = foundCustomer.getRentals() ;
-		customerRentals.add(rental);
-		foundCustomer.setRentals(customerRentals);
-	}
-
-	public void register(String object) {
-		if ( object.equals("customer") ) {
-			System.out.println("Enter customer name: ") ;
-			String name = scanner.next();
-			Customer customer = new Customer(name) ;
-			customers.add(customer) ;
-		}
-		else {
-			System.out.println("Enter video title to register: ") ;
-			String title = scanner.next() ;
-
-			System.out.println("Enter video type( 1 for VHD, 2 for CD, 3 for DVD ):") ;
-			int videoType = scanner.nextInt();
-
-			System.out.println("Enter price code( 1 for Regular, 2 for New Release ):") ;
-			int priceCode = scanner.nextInt();
-
-			Date registeredDate = new Date();
-			Video video = new Video(title, videoType, priceCode, registeredDate) ;
-			videos.add(video) ;
-		}
-	}
-
-	public int showCommand() {
+	public void showCommand() {
 		System.out.println("\nSelect a command !");
 		System.out.println("\t 0. Quit");
 		System.out.println("\t 1. List customers");
@@ -215,10 +35,48 @@ public class VRUI {
 		System.out.println("\t 6. Return video");
 		System.out.println("\t 7. Show customer report");
 		System.out.println("\t 8. Show customer and clear rentals");
-
+	}
+	public int inputCommand() {
 		int command = scanner.nextInt() ;
-
 		return command ;
+	}
 
+	public boolean isQuit(int command) {
+		return (command == 0);
+	}
+
+	private String inputCustomer() {
+		System.out.println("Enter customer name: ") ;
+		String customerName = scanner.next() ;
+		return customerName;
+	}
+
+	public void processCommand(int command) {
+		String customerName;
+		switch ( command ) {
+			//case 0: quit = true ; break ;
+			case 1: rentalService.listCustomers() ; break ;
+			case 2: rentalService.listVideos() ; break ;
+			case 3: rentalService.register("customer") ; break ;
+			case 4: rentalService.register("video") ; break ;
+			case 5:
+				customerName = inputCustomer();
+				rentalService.rentVideo(customerName);
+				break;
+			case 6:
+				customerName = inputCustomer();
+				rentalService.returnVideo(customerName);
+				break;
+			case 7:
+				customerName = inputCustomer();
+				rentalService.getCustomerReport(customerName);
+				break;
+			case 8:
+				customerName = inputCustomer();
+				rentalService.clearRentals(customerName);
+				break ;
+			case -1: rentalService.init() ; break ;
+			default: break ;
+		}
 	}
 }
